requirements.txt
alembic==1.7.7
anyio==3.5.0
asgiref==3.5.0
attrs==21.4.0
automium==0.2.6
automium-web==0.1.1
cffi==1.15.0
click==8.0.4
cryptography==36.0.2
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.75.0
Flask==2.1.1
functions==0.7.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
iniconfig==1.1.1
install==1.3.5
itsdangerous==2.1.1
Jinja2==3.0.3
legacy==0.1.6
Mako==1.2.0
MarkupSafe==2.1.1
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.8
pytest==7.1.1
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.8
schemas==0.7.1
six==1.16.0
sniffio==1.2.0
sqladmin==0.1.7
SQLAlchemy==1.4.32
starlette==0.17.1
Starlette-WTF==0.4.3
tomli==2.0.1
typing_extensions==4.1.1
uvicorn==0.17.6
Werkzeug==2.1.1
WTForms==3.0.1

motor~=3.0.0
pymongo~=4.1.1